refactor drawManager.js
